+++ app/controllers/management/datasets_controller.rb
@@ -9,18 +9,7 @@ class Management::DatasetsController < ManagementController
   before_action :authenticate_user_for_site!
 
   def index
-    metadata_list = DatasetService.metadata_find_by_ids(session[:user_token], @datasets.map(&:id))
-    @dataset_metadata =
-      Hash[metadata_list.map do |d|
-        attributes = d['attributes'].symbolize_keys
-        application_properties = attributes[:applicationProperties].try(:symbolize_keys)
-        metadata = attributes.slice(*Dataset::API_PROPERTIES)
-        metadata = metadata.merge(application_properties.slice(*Dataset::APPLICATION_PROPERTIES)) if application_properties.present?
-        [
-          attributes[:dataset],
-          metadata.values.reject{ |v| v.blank? }
-        ]
-      end]
+    @dataset_metadata = Dataset.get_metadata_list_for_frontend(session[:user_token], @datasets.map(&:id))
 
     gon.datasets = @datasets.map do |dataset|
       {