+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+- Added metadata table modal to management section.
+- Added metadata table modal to open content widgets.
+- Added metadata table modal to analysis dashboards.
+
+### 25 April 2017
 - Add metadata fields to dataset creation and list
 - Improved 404 page
 - Added sitemap.xml and sitemap.html pages